requirements.txt
streamlit
torch
transformers
sentence-transformers
faiss-cpu
pandas
numpy
scikit-learn
tqdm
requests
